# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API DoReMi - Plateforme scolaire numérique",
        "description": "Bienvenue dans la documentation de l'API **DoReMi**, une plateforme complète de gestion scolaire hybride.\\n\\n\n        ## Modules couverts :\\n\n        - Gestion des écoles et utilisateurs (élèves, enseignants, parents)\\n\n        - Cours en ligne et en présentiel\\n\n        - Suivi des notes et bulletins\\n\n        - Gestion des documents et devoirs\\n\n        - Actualités, emploi du temps, messagerie\\n\n        - Gestion des stages : demandes & offres\\n\\n\n        Cette documentation est à jour pour l'environnement de développement.",
        "termsOfService": "https://doremi-app.test/conditions",
        "contact": {
            "name": "Support Technique DoReMi",
            "email": "support@doremi-app.test"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/api/internships": {
            "get": {
                "tags": [
                    "Stages"
                ],
                "summary": "Lister tous les stages",
                "operationId": "e74ba8f8c38595f4ef3ae05c4cbe1390",
                "responses": {
                    "200": {
                        "description": "Liste des stages"
                    }
                }
            },
            "post": {
                "tags": [
                    "Stages"
                ],
                "summary": "Créer un nouveau stage",
                "operationId": "2a7b395aba72f9bf95f5530b3318e6d8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "company",
                                    "description",
                                    "location",
                                    "recruiter_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Stage Développement Web"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Société Tech"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Description du stage"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Dakar"
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "6 mois"
                                    },
                                    "requirements": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "application_deadline": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "salary": {
                                        "type": "string",
                                        "example": "150000 FCFA"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "stage",
                                            "alternance",
                                            "emploi"
                                        ]
                                    },
                                    "remote": {
                                        "type": "boolean"
                                    },
                                    "is_premium": {
                                        "type": "boolean"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "recruiter_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stage créé avec succès"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/internships/{id}": {
            "get": {
                "tags": [
                    "Stages"
                ],
                "summary": "Afficher un stage spécifique",
                "operationId": "34ebc6dd4c3fe8e33b4a77233dcf1324",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du stage",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails du stage"
                    },
                    "404": {
                        "description": "Stage non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Stages"
                ],
                "summary": "Mettre à jour un stage",
                "operationId": "dc03c095549bafd1e28057eb53b4a974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du stage à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "company": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "location": {
                                        "type": "string"
                                    },
                                    "duration": {
                                        "type": "string"
                                    },
                                    "requirements": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "application_deadline": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "salary": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "remote": {
                                        "type": "boolean"
                                    },
                                    "is_premium": {
                                        "type": "boolean"
                                    },
                                    "tags": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "status": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Stage mis à jour avec succès"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Stages"
                ],
                "summary": "Supprimer un stage",
                "operationId": "1ce8a406e4c5fdcce2956d6e4213e790",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du stage à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stage supprimé avec succès"
                    },
                    "404": {
                        "description": "Stage non trouvé"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/internships/my-offers": {
            "get": {
                "tags": [
                    "Stages"
                ],
                "summary": "Lister mes offres de stage (pour les recruteurs)",
                "operationId": "62077df924909e6ebf1517f81127275e",
                "responses": {
                    "200": {
                        "description": "Liste des offres du recruteur connecté"
                    },
                    "403": {
                        "description": "Accès refusé - Seuls les recruteurs peuvent accéder à cette ressource"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/internships/search": {
            "get": {
                "tags": [
                    "Stages"
                ],
                "summary": "Rechercher des stages",
                "operationId": "1d34a50ae1db9adee3625390cb4ee816",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Terme de recherche",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Type de stage",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "stage",
                                "alternance",
                                "emploi"
                            ]
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Localisation",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "remote",
                        "in": "query",
                        "description": "Télétravail possible",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats de la recherche"
                    }
                }
            }
        },
...
    },
...
}